backend/requirements.txt
Django==2.2.19
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.0.4
isort==5.10.1
psycopg2-binary==2.8.6
python-dotenv==0.20.0
